+++ CHANGES.rst
@@ -17,6 +17,7 @@ Unreleased
         ``key in data`` instead.
     -   ``Request.disable_data_descriptor`` is removed, pass
         ``shallow=True`` instead.
+    -   Remove the ``no_etag`` parameter from ``Response.freeze()``.
 
 -   Default values passed to ``Headers`` are validated the same way
     values added later are. :issue:`1608`